requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-binary==104.0.5112.79.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
Flask==2.2.1
fonttools==4.34.4
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
MouseInfo==0.1.3
nltk==3.7
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.6.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
regex==2022.7.25
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
telebot==0.0.4
threadpoolctl==3.1.0
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
Werkzeug==2.2.1
wikipedia==1.4.0
wsproto==1.1.0
XLML==0.1.2
